refactor: unify executor abstraction and add CI clippy checks
- Make NoopExecutor platform-agnostic (works on native and WASM)
- Merge BlockingExecutor into NoopExecutor with deprecated aliases
- Update fusio-parquet to use Executor trait with NoopExecutor default
- Update fusio-manifest to use NoopExecutor instead of BlockingExecutor
- Add clippy checks to CI for all workspace members (native + WASM)
- Fix clippy warnings in OPFS code
- Rename wasm-http feature to web-http with backward-compat alias

+++ fusio/src/impls/disk/opfs/mod.rs
@@ -76,7 +76,7 @@ impl OPFSFile {
         let file_handle = self.file_handle.as_ref().expect("read file after closed.");
         let file = promise::<File>(file_handle.get_file()).await?;
 
-        if (file.size().round() as u64) < pos + buf_len as u64 {
+        if (file.size().round() as u64) < pos + buf_len {
             return Err(Error::Io(io::Error::new(
                 io::ErrorKind::UnexpectedEof,
                 "Read unexpected eof",
